Add sort command
Address book does not sort the contacts added to the list automatically.

Searching through contacts can get troublesome especially if the user
is unable to fully recall the name of the individual he is trying to
find.

Let's add a feature to sort the contacts in the list by their names
in alphabetical order.

Alphabetical ordering of names is a very intuitive way of sorting
and would help the user navigate his contacts easier.